+++ Code2Gether-Discord-Bot/Modules/PrivilegedModules/MakeChannelModule.cs
@@ -1,10 +1,9 @@
 ﻿using System.Threading.Tasks;
 using Code2Gether_Discord_Bot.Static;
 using Discord;
-using Discord.Audio.Streams;
 using Discord.Commands;
 
-namespace Code2Gether_Discord_Bot.Modules
+namespace Code2Gether_Discord_Bot.Modules.PrivilegedModules
 {
     public class MakeChannelModule : ModuleBase<SocketCommandContext>
     {
