+++ Code2Gether-Discord-Bot/Modules/UserModules/PingModule.cs
@@ -1,12 +1,8 @@
-﻿using Code2Gether_Discord_Bot.Static;
-using Discord;
+﻿using System.Threading.Tasks;
+using Code2Gether_Discord_Bot.Static;
 using Discord.Commands;
-using System;
-using System.Collections.Generic;
-using System.Text;
-using System.Threading.Tasks;
 
-namespace Code2Gether_Discord_Bot.Modules
+namespace Code2Gether_Discord_Bot.Modules.UserModules
 {
     public class PingModule : ModuleBase<SocketCommandContext>
     {
